backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4
twilio>=8.10.0
sendgrid>=6.10.0
python-multipart>=0.0.6
alembic>=1.12.1
python-dotenv>=1.0.0
bcrypt>=4.1.1
pytest>=7.4.3
httpx>=0.25.2
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0.post1